Request RECORD_AUDIO at startup instead of from XServerScreen
Requesting it from XServerScreen raised a system dialog at the worst
possible moment. The LaunchedEffect fired on screen entry, while the
container was still booting and before the AndroidView factory had
created xServerView, and the dialog backgrounds the activity for as long
as it is up. Two consequences:

- setupXEnvironment finishes with isActivityInForeground false, so the
  branch right after it suspends the environment immediately; on the
  manual-resume suspend policy the user has to resume a game they just
  launched.
- onActivityDestroyed calls exit(xServerView!!...) with no null guard,
  and it is registered during composition, well before xServerView is
  assigned. Anything that destroys the activity while the dialog is up
  (don't-keep-activities, memory pressure, a configuration change) lands
  on that non-null assertion. The sharp edge already existed; putting a
  dialog there made it far easier to reach.

Moving the request next to the existing POST_NOTIFICATIONS one resolves
the permission before any container exists, so default.pa is written
correctly at daemon spawn and neither problem is reachable. The two are
sequenced rather than fired together, since the platform drops a
permission request made while another dialog is showing.

The trade-off is that the prompt no longer carries launch context and is
shown to users who never touch voice chat. Reported by @joshuatam.
