+++ app/src/main/java/app/gamenative/ui/screen/xserver/XServerScreen.kt
@@ -1,9 +1,7 @@
 package app.gamenative.ui.screen.xserver
 
-import android.Manifest
 import android.app.Activity
 import android.content.Context
-import android.content.pm.PackageManager
 import android.database.ContentObserver
 import android.graphics.Color
 import android.os.Build
@@ -27,8 +25,6 @@ import android.hardware.input.InputManager
 import android.view.InputDevice
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.BackHandler
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
 import app.gamenative.BuildConfig
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
@@ -388,34 +384,6 @@ fun XServerScreen(
         ContainerUtils.getContainer(context, appId)
     }
     val activity = remember(context) { BrightnessManager.findActivity(context) }
-
-    // Games that open a recording device otherwise get AAudioSink.monitor, i.e. a loopback
-    // of their own output. Ask for the microphone here rather than at app startup so the
-    // prompt has context, and only for containers that actually use the pulseaudio driver.
-    // If it is denied, PulseAudioComponent simply starts without a capture source.
-    val micPermissionLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.RequestPermission(),
-    ) { isGranted ->
-        Timber.i("RECORD_AUDIO permission granted: $isGranted")
-        if (isGranted) {
-            // The daemon reads default.pa once at spawn, and the container is usually
-            // already booting by the time this is answered, so load the capture source
-            // into the running daemon rather than making the user relaunch the game.
-            PluviaApp.xEnvironment
-                ?.getComponent(PulseAudioComponent::class.java)
-                ?.enableMicrophone()
-        }
-    }
-
-    LaunchedEffect(appId) {
-        if (!BuildConfig.MODERN_XR &&
-            container?.audioDriver == "pulseaudio" &&
-            ContextCompat.checkSelfPermission(context, Manifest.permission.RECORD_AUDIO) !=
-            PackageManager.PERMISSION_GRANTED
-        ) {
-            micPermissionLauncher.launch(Manifest.permission.RECORD_AUDIO)
-        }
-    }
 
     DisposableEffect(activity) {
         if (activity == null) return@DisposableEffect onDispose { }
